Update pull_all_apps to use regex for check
pull_all_apps calls a "git status" and checks if the response contains "nothing to commit, working directory clean". However, it seems that at least some versions of git output the message "nothing to commit, working directory clean". This causes bench update to not work in systems where the output message uses "tree" instead of "directory". This update is to allow both strings to be used.