restart supervisor on boot
This is being added to restart supervisor inside the virtual machine on
boot, since the command 'bench update' breaks on boot, since the
supervisor is not active. This is a bug in debian, which has been
propogated to ubuntu, and this piece of code can be removed when the bug
is fixed.