ci: Remove install script tests
The script has been untested and unmaintained for a while. We've asked
for help in maintenance but nobody seems to want to undertake it.

With newer versions of distros and packages coming out each day, setups
get more and more volatile. Making these tests run in the CI don't
reflect how it would go down irl as no server comes with the same base
config. Apart from the base sanity, these tests are just a waste of
carbon. Removing them.