fix(nginx): set `proxy_http_version` to `1.1`
This sets the `proxy_http_version` to `1.1`.

The reason this change should be implemented is that when this is deployed in a Kubernetes environment with Istio, the proxy request fails with HTTP error 426 as the HTTP version is too low for envoy to handle. Setting it to `1.1` will solve the issue.